+++ NEWS.md
@@ -6,3 +6,4 @@ code from the old SurveySolutionAPI package.
 * extended suso_getINT, such that if no sv_id is provided, it will return all interviewers in the workspace.
 * suso_PwCheck now prints a message and the url to the console when in interactive mode
 * suso_getINT_info now covers both interviewer details and retrieval of log files (log = T)
+* workspace has been added to suso_set_key